rename centernet_backbone.py to default_backbone.py
